requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
numpy==2.1.3
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
apache-airflow[postgres,google,slack,celery]
psycopg2-binary==2.9.7
pendulum==2.1.2
flask-session==0.4.0

